+++ HW2/server.py
@@ -1,7 +1,8 @@
-from socket_stream import SocketStream 
+from file_transfer import FileTransfer
 import logging
 
-from config import *
+from config import LOGGING_FORMAT
+
 
 logging.basicConfig(level=logging.INFO, 
                     format=LOGGING_FORMAT,
@@ -9,10 +10,11 @@ logging.basicConfig(level=logging.INFO,
                     filemode="w")
 logger = logging.getLogger(__name__)
 
-class FileTranferServer(SocketStream):
+
+class FileTranferServer(FileTransfer):
     def __init__(self, host: str, port: int):
         super().__init__(host, port,
-                      is_server=True,
-                      base_logger=logger)
+                         is_server=True,
+                         base_logger=logger)
