+++ HW2/socket_stream.py
@@ -1,6 +1,7 @@
 import socket
 import logging
 
+
 class SocketStream:
     def __init__(self, host: str, port: int, 
                  is_server: bool = False,
@@ -73,21 +74,3 @@ class SocketStream:
         data = conn.recv(bufsize)
         self.logger.info(f"Recived: {data}")
         return data
-
-    def send_file(self, filepath: str, bufsize: int):
-        self.logger.info(f"Starting sending file: {filepath}")
-        with open(filepath, "rb") as f:
-            while chunk := f.read(bufsize):
-                self.sendb(chunk)
-                self.logger.info(f"Sended: {len(chunk)} bytes of file")
-        self.logger.info(f"Ended sending file: {filepath}")
-
-    def recv_file(self, filepath: str, bufsize: int):
-        self.logger.info(f"Starting reciving file: {filepath}")
-        with open(filepath, "wb") as f:
-            while chunk := self.recvb(bufsize):
-                f.write(chunk)
-                self.logger.info(f"Recived: {len(chunk)} bytes of file")
-        self.logger.info(f"Ended reciving file: {filepath}")
-
-
